move PG to Binary
