+++ squeal-postgresql/src/Squeal/PostgreSQL/Manipulation.hs
@@ -62,6 +62,7 @@ import GHC.TypeLits (KnownSymbol)
 import qualified Generics.SOP as SOP
 import qualified GHC.Generics as GHC
 
+import Squeal.PostgreSQL.Binary
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Render
 import Squeal.PostgreSQL.Query
