+++ squeal-postgresql/src/Squeal/PostgreSQL/Schema.hs
@@ -40,11 +40,6 @@ module Squeal.PostgreSQL.Schema
   , NullityType (..)
   , RowType
   , FromType
-    -- * Haskell to Postgres Types
-  , PG
-  , NullPG
-  , TuplePG
-  , RowPG
     -- * Schema Types
   , ColumnType
   , ColumnsType
@@ -78,7 +73,6 @@ module Squeal.PostgreSQL.Schema
     -- * Enumerated Labels
   , IsPGlabel (..)
   , PGlabel (..)
-  , LabelsPG
     -- * Grouping
   , Grouping (..)
   , GroupedBy
@@ -123,29 +117,18 @@ module Squeal.PostgreSQL.Schema
 
 import Control.Category
 import Control.DeepSeq
-import Data.Aeson (Value)
 import Data.ByteString (ByteString)
-import Data.Int (Int16, Int32, Int64)
 import Data.Kind
 import Data.Monoid hiding (All)
-import Data.Scientific (Scientific)
 import Data.String
-import Data.Text (Text)
-import Data.Time
-import Data.Word (Word16, Word32, Word64)
 import Data.Type.Bool
-import Data.UUID.Types (UUID)
+import Data.Word (Word32)
 import Generics.SOP
-import Generics.SOP.Record
 import GHC.OverloadedLabels
 import GHC.TypeLits
-import Network.IP.Addr
 import Prelude hiding (id, (.))
 
-import qualified Data.ByteString.Lazy as Lazy
-import qualified Data.Text.Lazy as Lazy
 import qualified GHC.Generics as GHC
-import qualified Generics.SOP.Type.Metadata as Type
 
 import Squeal.PostgreSQL.Render
 
@@ -773,151 +756,6 @@ instance All KnownSymbol labels => RenderSQL (NP PGlabel labels) where
     = commaSeparated
     . hcollapse
     . hcmap (Proxy @KnownSymbol) (K . renderSQL)
-
-{- | The `PG` type family embeds a subset of Haskell types
-as Postgres types. As an open type family, `PG` is extensible.
-
->>> :kind! PG LocalTime
-PG LocalTime :: PGType
-= 'PGtimestamp
-
->>> newtype MyDouble = My Double
->>> type instance PG MyDouble = 'PGfloat8
--}
-type family PG (hask :: Type) :: PGType
-type instance PG Bool = 'PGbool
-type instance PG Int16 = 'PGint2
-type instance PG Int32 = 'PGint4
-type instance PG Int64 = 'PGint8
-type instance PG Word16 = 'PGint2
-type instance PG Word32 = 'PGint4
-type instance PG Word64 = 'PGint8
-type instance PG Scientific = 'PGnumeric
-type instance PG Float = 'PGfloat4
-type instance PG Double = 'PGfloat8
-type instance PG Char = 'PGchar 1
-type instance PG Text = 'PGtext
-type instance PG Lazy.Text = 'PGtext
-type instance PG String = 'PGtext
-type instance PG ByteString = 'PGbytea
-type instance PG Lazy.ByteString = 'PGbytea
-type instance PG LocalTime = 'PGtimestamp
-type instance PG UTCTime = 'PGtimestamptz
-type instance PG Day = 'PGdate
-type instance PG TimeOfDay = 'PGtime
-type instance PG (TimeOfDay, TimeZone) = 'PGtimetz
-type instance PG DiffTime = 'PGinterval
-type instance PG UUID = 'PGuuid
-type instance PG (NetAddr IP) = 'PGinet
-type instance PG Value = 'PGjson
-
-{-| The `LabelsPG` type family calculates the constructors of a
-Haskell enum type.
-
->>> data Schwarma = Beef | Lamb | Chicken deriving GHC.Generic
->>> instance Generic Schwarma
->>> instance HasDatatypeInfo Schwarma
->>> :kind! LabelsPG Schwarma
-LabelsPG Schwarma :: [Type.ConstructorName]
-= '["Beef", "Lamb", "Chicken"]
--}
-type family LabelsPG (hask :: Type) :: [Type.ConstructorName] where
-  LabelsPG hask =
-    ConstructorNamesOf (ConstructorsOf (DatatypeInfoOf hask))
-
-{- | `RowPG` turns a Haskell record type into a `RowType`.
-
->>> data Person = Person { name :: Text, age :: Int32 } deriving GHC.Generic
->>> instance Generic Person
->>> instance HasDatatypeInfo Person
->>> :kind! RowPG Person
-RowPG Person :: [(Symbol, NullityType)]
-= '["name" ::: 'NotNull 'PGtext, "age" ::: 'NotNull 'PGint4]
--}
-type family RowPG (hask :: Type) :: RowType where
-  RowPG (hask1, hask2) = Join (RowPG hask1) (RowPG hask2)
-  RowPG (hask1, hask2, hask3) =
-    Join (RowPG hask1) (RowPG (hask2, hask3))
-  RowPG (hask1, hask2, hask3, hask4) =
-    Join (RowPG hask1) (RowPG (hask2, hask3, hask4))
-  RowPG (hask1, hask2, hask3, hask4, hask5) =
-    Join (RowPG hask1) (RowPG (hask2, hask3, hask4, hask5))
-  RowPG (P (col ::: head)) = '[col ::: NullPG head]
-  RowPG hask = RowOf (RecordCodeOf hask)
-
-type family RowOf (fields :: [(Symbol, Type)]) :: RowType where
-  RowOf '[] = '[]
-  RowOf (field ': fields) = FieldPG field ': RowOf fields
-
-type family FieldPG (field :: (Symbol, Type)) :: (Symbol, NullityType) where
-  FieldPG (field ::: hask) = field ::: NullPG hask
-
-{- | `NullPG` turns a Haskell type into a `NullityType`.
-
->>> :kind! NullPG Double
-NullPG Double :: NullityType
-= 'NotNull 'PGfloat8
->>> :kind! NullPG (Maybe Double)
-NullPG (Maybe Double) :: NullityType
-= 'Null 'PGfloat8
--}
-type family NullPG (hask :: Type) :: NullityType where
-  NullPG (Maybe hask) = 'Null (PG hask)
-  NullPG hask = 'NotNull (PG hask)
-
-{- | `TuplePG` turns a Haskell tuple type (including record types) into
-the corresponding list of `NullityType`s.
-
->>> :kind! TuplePG (Double, Maybe Char)
-TuplePG (Double, Maybe Char) :: [NullityType]
-= '[ 'NotNull 'PGfloat8, 'Null ('PGchar 1)]
--}
-type family TuplePG (hask :: Type) :: [NullityType] where
-  TuplePG hask = TupleOf (TupleCodeOf hask (Code hask))
-
-type family TupleOf (tuple :: [Type]) :: [NullityType] where
-  TupleOf '[] = '[]
-  TupleOf (hask ': tuple) = NullPG hask ': TupleOf tuple
-
-type family TupleCodeOf (hask :: Type) (code :: [[Type]]) :: [Type] where
-  TupleCodeOf hask '[tuple] = tuple
-  TupleCodeOf hask '[] =
-    TypeError
-      (    'Text "The type `" :<>: 'ShowType hask :<>: 'Text "' is not a tuple type."
-      :$$: 'Text "It is a void type with no constructors."
-      )
-  TupleCodeOf hask (_ ': _ ': _) =
-    TypeError
-      (    'Text "The type `" :<>: 'ShowType hask :<>: 'Text "' is not a tuple type."
-      :$$: 'Text "It is a sum type with more than one constructor."
-      )
-
--- | Calculates constructors of a datatype.
-type family ConstructorsOf (datatype :: Type.DatatypeInfo)
-  :: [Type.ConstructorInfo] where
-    ConstructorsOf ('Type.ADT _module _datatype constructors) =
-      constructors
-    ConstructorsOf ('Type.Newtype _module _datatype constructor) =
-      '[constructor]
-
--- | Calculates the name of a nullary constructor, otherwise
--- generates a type error.
-type family ConstructorNameOf (constructors :: Type.ConstructorInfo)
-  :: Type.ConstructorName where
-    ConstructorNameOf ('Type.Constructor name) = name
-    ConstructorNameOf ('Type.Infix name _assoc _fix) = TypeError
-      ('Text "ConstructorNameOf error: non-nullary constructor "
-        ':<>: 'Text name)
-    ConstructorNameOf ('Type.Record name _fields) = TypeError
-      ('Text "ConstructorNameOf error: non-nullary constructor "
-        ':<>: 'Text name)
-
--- | Calculate the names of nullary constructors.
-type family ConstructorNamesOf (constructors :: [Type.ConstructorInfo])
-  :: [Type.ConstructorName] where
-    ConstructorNamesOf '[] = '[]
-    ConstructorNamesOf (constructor ': constructors) =
-      ConstructorNameOf constructor ': ConstructorNamesOf constructors
 
 -- | Is a type a valid JSON key?
 type PGJsonKey = '[ 'PGint2, 'PGint4, 'PGtext ]
